move ref to elementid to DataFormats/Common
